refactor: id Long 타입으로 변경.

+++ src/main/java/io/cavia/trader/module/member/repository/MemberMapper.java
@@ -50,7 +50,7 @@ public interface MemberMapper {
     /**
      * 회원을 삭제합니다
      */
-    int delete(int id);
+    int delete(Long id);
 
     /**
      * 이메일이 이미 있는지 조회합니다.
@@ -71,5 +71,5 @@ public interface MemberMapper {
     /**
      * 비밀번호가 있는지 조회.
      */
-    boolean existsByIdAndPassword(int id, String password);
+    boolean existsByIdAndPassword(Long id, String password);
 }
