+++ src/main/java/io/cavia/trader/module/member/repository/MemberRepository.java
@@ -40,7 +40,7 @@ public interface MemberRepository {
     /**
      * 회원을 삭제합니다
      */
-    int delete(int id);
+    int delete(Long id);
 
     /**
      * 이메일이 이미 있는지 조회합니다.
@@ -61,7 +61,7 @@ public interface MemberRepository {
     /**
      * 비밀번호가 있는지 조회.
      */
-    boolean existsByIdAndPassword(int id, String password);
+    boolean existsByIdAndPassword(Long id, String password);
 
     /**
      * id로 멤버를 조회함
